requirements.txt
nbconvert
langchain-cli
langchain
langchain-openai
langserve[all]
python-dotenv==1.0.0
langchain-qdrant
qdrant_client